add more tests for lstm
