+++ dl/src/main/scala/com/intel/analytics/bigdl/nn/LSTMPeephole.scala
@@ -17,7 +17,6 @@
 
 package com.intel.analytics.bigdl.nn
 
-import com.intel.analytics.bigdl._
 import com.intel.analytics.bigdl.nn.abstractnn.AbstractModule
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
@@ -25,7 +24,7 @@ import com.intel.analytics.bigdl.utils.{T, Table}
 
 import scala.reflect.ClassTag
 
-class LSTMCell[T : ClassTag] (
+class LSTMPeephole[T : ClassTag] (
   inputSize: Int = 4,
   hiddenSize: Int = 3,
   p: Double = 0)
@@ -40,17 +39,19 @@ class LSTMCell[T : ClassTag] (
 
   def buildGate(): Sequential[T] = {
     val gate = Sequential()
-      .add(NarrowTable(1, 2))
 
     val i2g = Sequential()
+      .add(Dropout(p))
       .add(Linear(inputSize, hiddenSize))
     val h2g = Sequential()
-      .add(Linear(hiddenSize, hiddenSize))
+      .add(Dropout(p))
+      .add(Linear(hiddenSize, hiddenSize, withBias = false))
 
     gate
       .add(ParallelTable()
         .add(i2g)
-        .add(h2g))
+        .add(h2g)
+        .add(CMul(Array(hiddenSize))))
       .add(CAddTable())
       .add(Sigmoid())
   }
@@ -75,9 +76,11 @@ class LSTMCell[T : ClassTag] (
       .add(NarrowTable(1, 2))
 
     val i2h = Sequential()
+      .add(Dropout(p))
       .add(Linear(inputSize, hiddenSize))
     val h2h = Sequential()
-      .add(Linear(hiddenSize, hiddenSize))
+      .add(Dropout(p))
+      .add(Linear(hiddenSize, hiddenSize, withBias = false))
 
     hidden
       .add(ParallelTable()
@@ -172,163 +175,13 @@ class LSTMCell[T : ClassTag] (
   }
 }
 
-object LSTMCell {
-  def apply[@specialized(Float, Double) T: ClassTag](
-    inputSize: Int = 4,
-    hiddenSize: Int = 3)
-    (implicit ev: TensorNumeric[T]): LSTMCell[T] = {
-    new LSTMCell[T](inputSize, hiddenSize)
-  }
-}
-
-class FastLSTMCell[T : ClassTag] (
-  inputSize: Int = 4,
-  hiddenSize: Int = 3,
-  p: Double = 0,
-  private var initMethod: InitializationMethod = Default)
-  (implicit ev: TensorNumeric[T])
-  extends AbstractModule[Table, Table, T] {
-  var gates: Sequential[T] = _
-  var cellLayer: Sequential[T] = _
-  var lstm: Sequential[T] = buildLSTM()
-
-  def buildGates(): Sequential[T] = {
-    val gates = Sequential()
-      .add(NarrowTable(2, 2))
-
-    var i2g: AbstractModule[_, _, T] = null
-    var h2g: AbstractModule[_, _, T] = null
-
-    if (p != 0) {
-      i2g = Sequential()
-        .add(ConcatTable()
-          .add(Dropout(p))
-          .add(Dropout(p))
-          .add(Dropout(p))
-          .add(Dropout(p)))
-        .add(ParallelTable()
-          .add(Linear(inputSize, hiddenSize))
-          .add(Linear(inputSize, hiddenSize))
-          .add(Linear(inputSize, hiddenSize))
-          .add(Linear(inputSize, hiddenSize)))
-        .add(JoinTable(2, 1))
-
-      val o2g = Sequential()
-        .add(ConcatTable()
-          .add(Dropout(p))
-          .add(Dropout(p))
-          .add(Dropout(p))
-          .add(Dropout(p)))
-        .add(ParallelTable()
-          .add(Linear(hiddenSize, hiddenSize, withBias = false))
-          .add(Linear(hiddenSize, hiddenSize, withBias = false))
-          .add(Linear(hiddenSize, hiddenSize, withBias = false))
-          .add(Linear(hiddenSize, hiddenSize, withBias = false)))
-        .add(JoinTable(2, 1))
-    } else {
-      i2g = Linear(inputSize, 4 * hiddenSize)
-      h2g = Linear(hiddenSize, 4 * hiddenSize, withBias = false)
-    }
-
-    gates
-      .add(ParallelTable()
-        .add(h2g)
-        .add(i2g))
-      .add(CAddTable())
-      .add(Reshape(Array(4, hiddenSize)))
-      .add(SplitTable(1, 2))
-      .add(ParallelTable()
-        .add(Sigmoid())
-        .add(Tanh())
-        .add(Sigmoid())
-        .add(Sigmoid()))
-
-    this.gates = gates
-    gates
-  }
-
-  def buildLSTM(): Sequential[T] = {
-    buildGates()
-
-    val lstm = Sequential()
-      .add(FlattenTable())
-      .add(ConcatTable()
-        .add(SelectTable(1))
-        .add(gates))
-      .add(FlattenTable()) // cell, input, hidden, forget, output
-
-    val cellLayer = Sequential()
-      .add(ConcatTable()
-        .add(Sequential()
-          .add(NarrowTable(2, 2))
-          .add(CMulTable()))
-        .add(Sequential()
-          .add(ConcatTable()
-            .add(SelectTable(1))
-            .add(SelectTable(4)))
-          .add(CMulTable())))
-      .add(CAddTable())
-
-    lstm
-      .add(ConcatTable()
-        .add(cellLayer)
-        .add(SelectTable(5)))
-      .add(FlattenTable())
-
-
-    lstm
-      .add(ConcatTable()
-        .add(SelectTable(1))
-        .add(Sequential()
-          .add(ConcatTable()
-            .add(Sequential()
-              .add(SelectTable(1))
-              .add(Tanh()))
-            .add(SelectTable(2)))
-          .add(CMulTable())))
-      .add(ConcatTable()
-        .add(Identity())
-        .add(SelectTable(2)))
-
-    output = T(T(), Tensor())
-    this.lstm = lstm
-    lstm
-  }
-
-
-  override def updateOutput(input: Table): Table = {
-    output = lstm.updateOutput(input).toTable
-    output
-  }
-
-  override def updateGradInput(input: Table, gradOutput: Table): Table = {
-    gradInput = lstm.updateGradInput(input, gradOutput).toTable
-    gradInput
-  }
-
-  override def accGradParameters(input: Table, gradOutput: Table, scale: Double): Unit = {
-    lstm.accGradParameters(input, gradOutput, scale)
-  }
-
-  override def updateParameters(learningRate: T): Unit = {
-    lstm.updateParameters(learningRate)
-  }
-
-  override def zeroGradParameters(): Unit = {
-    lstm.zeroGradParameters()
-  }
-
-  override def parameters(): (Array[Tensor[T]], Array[Tensor[T]]) = {
-    lstm.parameters()
-  }
-}
-
-object FastLSTMCell {
+object LSTMPeephole {
   def apply[@specialized(Float, Double) T: ClassTag](
     inputSize: Int = 4,
     hiddenSize: Int = 3,
     p: Double = 0)
-    (implicit ev: TensorNumeric[T]): FastLSTMCell[T] = {
-    new FastLSTMCell[T](inputSize, hiddenSize)
+    (implicit ev: TensorNumeric[T]): LSTMPeephole[T] = {
+    new LSTMPeephole[T](inputSize, hiddenSize, p)
   }
 }
+
